% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{extract_rank_step_names}
\alias{extract_rank_step_names}
\title{Extract the names of steps appearing once per rank}
\usage{
extract_rank_step_names(dx, id_tag = "start")
}
\arguments{
\item{dx}{an eventselection raw dataframe}

\item{id_tag}{the name of a step known to appear once per rank}
}
\value{
character vector containing the names
}
\description{
Extract the names of steps appearing once per rank
}
