% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{make_events_df}
\alias{make_events_df}
\title{Extract event-level data from an evenselection dataframe.}
\usage{
make_events_df(dx)
}
\arguments{
\item{dx}{an eventselection raw dataframe}
}
\value{
a tibble containing event-level data
}
\description{
Each row in the returned dataframe corresponds to a single event
processed.
}
