% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{make_rank_df}
\alias{make_rank_df}
\title{Create a dataframe of all the quantities recorded only once per rank.}
\usage{
make_rank_df(dx)
}
\arguments{
\item{dx}{an eventselection raw dataframe}
}
\value{
a tibble containing once-per-rank data
}
\description{
Create a dataframe of all the quantities recorded only once per rank.
}
