% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{raw_data_version}
\alias{raw_data_version}
\title{Return the version number (an integer) for the given (raw) datarame.}
\usage{
raw_data_version(df)
}
\arguments{
\item{df}{A dataframe, as obtained by read_csv from a raw timestamp file}
}
\value{
The version number of the data format
}
\description{
version 1 data files are recognized by having 3 columns
version 2 data files are recognized by having 4 columns
}
\details{
Any other number of columns produces an error.
}
