% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_dataframes.R
\name{read_raw_dataframes}
\alias{read_raw_dataframes}
\title{Read several raw dataframes}
\usage{
read_raw_dataframes(fileglob)
}
\arguments{
\item{fileglob}{a glob pattern identifying all the files to read}
}
\value{
a tibble combining all the dataframes read from the files
}
\description{
Read several raw dataframes
}
