% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_timestamp.R
\name{crop_timestamp}
\alias{crop_timestamp}
\title{Function to Crop the Timestamp Area from an Image}
\usage{
crop_timestamp(image_path, media_type = "image")
}
\arguments{
\item{image_path}{raw image path}

\item{media_type}{image or movie}
}
\description{
...
}
