% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{make_events_df}
\alias{make_events_df}
\title{Extract event-level data from an evenselection dataframe.}
\usage{
make_events_df(dx)
}
\arguments{
\item{dx}{an eventselection raw dataframe}
}
\value{
a tibble containing event-level data
}
\description{
Each row in the returned dataframe corresponds to a single event
processed.
}
\details{
The dataframe columns are:
\describe{
   \item{rank}{the MPI rank on which the call was made}
   \item{start}{time at start of the rank}
   \item{precr}{timestamp before create_records call}
   \item{postcr}{timestampe after create_records call}
   \item{postfr}{timestamp after records have been filled}
   \item{postps}{timestamp after slices have been filter for selection}
   \item{load}{duration of loading data from HEPnOS (ms)}
   \item{rec}{duration of record creation (ms)}
   \item{filt}{duration of filter criterion application (ms)}
   \item{evt}{id of this event}
   \item{bid}{id of block which arried this event}
   \item{nbytes}{sum of the size of records (in bytes) of this event}
   \item{nslices}{number of slices in this event}
}
}
