% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{make_global_df}
\alias{make_global_df}
\title{Extract global (once per rank) data from an eventselection dataframe}
\usage{
make_global_df(dx)
}
\arguments{
\item{dx}{an eventselection raw dataframe}
}
\value{
a tibble containing global data
}
\description{
Extract global (once per rank) data from an eventselection dataframe
}
