% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{make_reduction_loop1_df}
\alias{make_reduction_loop1_df}
\title{Extract reduction loop 1 (dequeue and merge) information from an eventselection dataframe}
\usage{
make_reduction_loop1_df(dx)
}
\arguments{
\item{dx}{an eventselection raw dataframe}
}
\value{
a tibble containing reduction loop 1 information
}
\description{
Extract reduction loop 1 (dequeue and merge) information from an eventselection dataframe
}
