% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{make_reduction_loop1_df}
\alias{make_reduction_loop1_df}
\title{Extract reduction loop 1 information from an eventselection dataframe.}
\usage{
make_reduction_loop1_df(dx)
}
\arguments{
\item{dx}{an eventselection raw dataframe}
}
\value{
a tibble containing reduction loop 1 information
}
\description{
Each row in the dataframe corresponds to an iteration in the
dequeue-and-merge loop. Note that some iterations may produce only a 'start_dequeue_loop'
record.
}
