% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{make_reduction_loop2_df}
\alias{make_reduction_loop2_df}
\title{Extract reduction loop 2 (enqueue) information from an eventselection
dataframe}
\usage{
make_reduction_loop2_df(dx)
}
\arguments{
\item{dx}{an eventselection raw dataframe}
}
\value{
a tibble containing reduction loop 2 information
}
\description{
Each row in the dataframe corresponds to an iteration in the enqueue loop.
}
\details{
The dataframe columns are:
\describe{
    \item{rank}{the MPI rank on which the loop was executed}
    \item{bid}{the id of the block on which reduceData was called}
    \item{round}{the reduction round}
    \item{idx}{the loop index for this iteration}
    \item{target_bid}{the id of the block to which the enqueued data are sent}
    \item{nenq}{number of slices enqueued this iteration}
    \item{start}{the time at the start of the loop}
    \item{end}{the time after the enqueuing}
    \item{t_tot}{total duration of the loop}
}
}
