% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventselection.R
\name{make_reduction_phase_df}
\alias{make_reduction_phase_df}
\title{Extract reduction round information from an eventselection dataframe.}
\usage{
make_reduction_phase_df(dx)
}
\arguments{
\item{dx}{an eventselection raw dataframe}
}
\value{
a tibble containing reduction phases information
}
\description{
Each row in the returned dataframe corresponds to one call to the reduceData.
function, for a given block in a rank.
}
