% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{read_raw}
\alias{read_raw}
\title{Create a tibble from a single rank timing file}
\usage{
read_raw(fname)
}
\arguments{
\item{fname}{name of the file to read}
}
\value{
a data.frame
}
\description{
Create a tibble from a single rank timing file
}
